[react-hw8] Refactor routing & fix loki tests
