+++ react-2021-09/game-of-life/README-hw08.md
@@ -56,14 +56,16 @@ All files                       |     100 |      100 |     100 |     100 |
   consts.ts                     |     100 |      100 |     100 |     100 |                   
  src/MyStorage                  |     100 |      100 |     100 |     100 |                   
   MyStorage.ts                  |     100 |      100 |     100 |     100 |                   
- src/components/App             |     100 |      100 |     100 |     100 |                   
-  App.tsx                       |     100 |      100 |     100 |     100 |                   
+ src/components/AppRouter       |     100 |      100 |     100 |     100 |                   
+  AppRouter.tsx                 |     100 |      100 |     100 |     100 |                   
  src/components/AppStateManager |     100 |      100 |     100 |     100 |                   
   AppStateManager.tsx           |     100 |      100 |     100 |     100 |                   
   appReducer.ts                 |     100 |      100 |     100 |     100 |                   
   playFieldUtils.ts             |     100 |      100 |     100 |     100 |                   
  src/components/AppStateView    |     100 |      100 |     100 |     100 |                   
   AppStateView.tsx              |     100 |      100 |     100 |     100 |                   
+ src/components/AppUI           |     100 |      100 |     100 |     100 |                   
+  AppUI.tsx                     |     100 |      100 |     100 |     100 |                   
  src/components/Button          |     100 |      100 |     100 |     100 |                   
   Button.tsx                    |     100 |      100 |     100 |     100 |                   
  src/components/Cell            |     100 |      100 |     100 |     100 |                   
@@ -87,7 +89,7 @@ All files                       |     100 |      100 |     100 |     100 |
 Test Suites: 11 passed, 11 total
 Tests:       64 passed, 64 total
 Snapshots:   0 total
-Time:        6.711 s
+Time:        6.422 s, estimated 7 s
 Ran all test suites.
