+++ react-2021-09/game-of-life/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "test": "jest",
     "test:loki": "npm run build-storybook && npx loki --requireReference --reactUri file:./storybook-static",
+    "check-ts": "npx tsc",
     "start": "webpack-dev-server --mode development",
     "build": "webpack",
     "lint": "npx eslint src/*.*",
