--- react-2021-09/game-of-life/src/components/App/index.tsx
@@ -1,1 +1,0 @@
-export * from './App';
